[Bugfix] Promote sparse-MLA KV row offsets to int64
The sparse-MLA Triton kernels compute `indices * stride_kv_token` in 32-bit.
A row id fits in int32, but the element offset does not once the flat KV view
passes 2**31 / stride rows: with the NoPE 512 layout that is 4,194,304 rows,
and with the 576 layout 3,728,271. Triton evaluates the whole offset
expression before the pointer addition widens anything, so the 64-bit base
pointer does not save it -- the multiply wraps first, the load lands on a
wrong cache row, and attention returns a plausible wrong answer with no
crash and no warning.

Six sites, all of the same shape:

  * _mla_sparse_kernel: the shared K/V gather, the separate-K gather, the
    RoPE tail gather and the separate-V gather;
  * the tuned Ampere prefill kernel's clamped row term, which feeds K, K_pe
    and V from one expression;
  * the gathered prefill MQA-logits kernel's `cols * D`, whose ceiling is
    16,777,216 gathered rows.

The decode paged MQA-logits kernel and the kpool cache writers already widen
their page indices before multiplying and are unchanged.

No behaviour change below the threshold: the widened product is equal to the
narrow one wherever the narrow one was correct.

+++ vllm/ampere_prefill/sparse_prefill_mla.py
@@ -204,7 +204,7 @@ def _sparse_mla_kernel(
         # Clamp instead of masking the gather: the rows of invalid entries are
         # read but their scores are forced to -1e30 below, so they contribute
         # exactly zero.  Row 0 is always in range whenever any row exists.
-        rows = tl.where(mask_kv, indices, 0) * stride_k_token \
+        rows = tl.where(mask_kv, indices, 0).to(tl.int64) * stride_k_token \
             + cur_kv_head_id * stride_k_head
 
         if V_IS_K:
